refactor: relocate app assembly to the shell and reduce gateway_oss to a facade
