+++ src/substack_gateway/app_factory.py
@@ -1,22 +1,26 @@
 from __future__ import annotations
 
-import contextlib
-import logging
-from collections.abc import AsyncIterator
-from contextlib import AsyncExitStack
-from typing import Any, cast
+from gateway_core.cache import configure_default_cache
 
-from starlette.applications import Starlette
-from starlette.middleware.cors import CORSMiddleware
-from starlette.responses import JSONResponse
-from starlette.routing import Mount, Route
+configure_default_cache()
 
-from gateway_oss.api.app import api
-from gateway_oss.application_features import build_oss_features
-from gateway_oss.extensions.base import ModuleInfo
-from gateway_oss.extensions.runtime import get_runtime
-from gateway_oss.mcp.app import mcp
-from gateway_oss.versioning import get_package_version
+import contextlib  # noqa: E402
+import logging  # noqa: E402
+from collections.abc import AsyncIterator  # noqa: E402
+from contextlib import AsyncExitStack  # noqa: E402
+from typing import Any, cast  # noqa: E402
+
+from starlette.applications import Starlette  # noqa: E402
+from starlette.middleware.cors import CORSMiddleware  # noqa: E402
+from starlette.responses import JSONResponse  # noqa: E402
+from starlette.routing import Mount, Route  # noqa: E402
+
+from gateway_oss.application_features import build_oss_features  # noqa: E402
+from gateway_oss.extensions.base import ModuleInfo  # noqa: E402
+from gateway_oss.versioning import get_package_version  # noqa: E402
+from substack_gateway.api_app import api  # noqa: E402
+from substack_gateway.mcp_app import mcp  # noqa: E402
+from substack_gateway.runtime import get_runtime  # noqa: E402
 
 _log = logging.getLogger(__name__)
 
@@ -24,7 +28,7 @@ APPLICATION_NAME = "substack-gateway"
 
 
 def _oss_module() -> ModuleInfo:
-    from gateway_oss.registry import load_mcp_capabilities, load_rest_capabilities
+    from substack_gateway.registry import load_mcp_capabilities, load_rest_capabilities
 
     registry_features: tuple[str, ...] = tuple(
         f
